requirements.txt
ipykernel
numpy
pandas
matplotlib
seaborn
flask
mlflow
scikit-learn
keras
tensorflow
hyperopt
dvc
dagshub
dvc_s3
boto3
pipreqs
apache-airflow
fastapi
uvicorn
pyd
pytest